+++ packages/metrics/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.3](/compare/@ceramicnetwork/metrics@0.0.3-rc.1...@ceramicnetwork/metrics@0.0.3) (2022-08-08)
+
+**Note:** Version bump only for package @ceramicnetwork/metrics
+
+
+
+
+
 ## [0.0.3-rc.1](/compare/@ceramicnetwork/metrics@0.0.3-rc.0...@ceramicnetwork/metrics@0.0.3-rc.1) (2022-07-14)
 
 **Note:** Version bump only for package @ceramicnetwork/metrics